third_party/WebKit/LayoutTests/fast/dom/shadow/get-element-by-id-in-shadow-root-expected.txt

CONSOLE WARNING: Calling Element.createShadowRoot() for an element which already hosts a shadow root is deprecated. See https://www.chromestatus.com/features/4668884095336448 for more details.
Tests to ensure that ShadowRoot.getElementById can get an element in ShadowRoot by its id. Can only run within DRT

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS getNodeInComposedTree('divA/inputB').id is "inputB"
PASS getNodeInComposedTree('divA/divC').id is "divC"
PASS getNodeInComposedTree('divA/divC/inputD').id is "inputD"
PASS getNodeInComposedTree('divA/divC//inputE').id is "inputE"
PASS getNodeInComposedTree('divA/divC///inputF').id is "inputF"
PASS successfullyParsed is true

TEST COMPLETE
